ritmo-backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pytest
pytest-asyncio
httpx